Move BindValidation tests to ValidationBindingTests.cs
