+++ src/ReactiveUI.Validation.Tests/PropertyValidationTests.cs
@@ -148,59 +148,6 @@ namespace ReactiveUI.Validation.Tests
             Assert.Equal("The value 'bon' is incorrect", validation.Text.ToSingleLine());
         }
 
-        /// <summary>
-        /// Using 2 validation rules ending with the same property name should not
-        /// result in both properties having all the errors of both properties.
-        /// </summary>
-        [Fact]
-        public void ErrorsWithTheSameLastPropertyShouldNotShareErrors()
-        {
-            var model = new SourceDestinationViewModel();
-            var view = new SourceDestinationView(model);
-
-            model.ValidationRule(
-                viewModel => viewModel.Source.Name,
-                name => !string.IsNullOrWhiteSpace(name),
-                "Source text");
-
-            model.ValidationRule(
-                viewModel => viewModel.Destination.Name,
-                name => !string.IsNullOrWhiteSpace(name),
-                "Destination text");
-
-            view.BindValidation(view.ViewModel, x => x.Source.Name, x => x.SourceError);
-            view.BindValidation(view.ViewModel, x => x.Destination.Name, x => x.DestinationError);
-
-            Assert.NotNull(view.SourceError);
-            Assert.Equal("Source text", view.SourceError);
-            Assert.Equal("Destination text", view.DestinationError);
-        }
-
-        /// <summary>
-        /// Property validations backed by ModelObservableValidationBase should
-        /// be bound to view as well as base property validations are.
-        /// </summary>
-        [Fact]
-        public void ComplexValidationRulesShouldBeBoundToView()
-        {
-            const string errorMessage = "Both inputs should be the same";
-            var view = new TestView(new TestViewModel
-            {
-                Name = "Josuke Hikashikata",
-                Name2 = "Jotaro Kujo"
-            });
-
-            view.ViewModel.ValidationRule(
-                m => m.Name,
-                m => m.WhenAnyValue(x => x.Name, x => x.Name2, (name, name2) => name == name2),
-                (vm, isValid) => isValid ? string.Empty : errorMessage);
-
-            view.BindValidation(view.ViewModel, x => x.Name, x => x.NameErrorLabel);
-
-            Assert.NotEmpty(view.NameErrorLabel);
-            Assert.Equal(errorMessage, view.NameErrorLabel);
-        }
-
         private static TestViewModel CreateDefaultValidModel()
         {
             return new TestViewModel { Name = "name" };
