+++ src/ReactiveUI.Validation.Tests/ValidationContextTests.cs
@@ -110,147 +110,6 @@ namespace ReactiveUI.Validation.Tests
         }
 
         /// <summary>
-        /// Verifies that two validations properties are correctly applied in a View property.
-        /// </summary>
-        [Fact]
-        public void ShouldSupportBindingTwoValidationsForOneProperty()
-        {
-            const int minimumLength = 5;
-
-            var viewModel = new TestViewModel { Name = "some" };
-            var view = new TestView(viewModel);
-
-            var firstValidation = new BasePropertyValidation<TestViewModel, string>(
-                viewModel,
-                vm => vm.Name,
-                s => !string.IsNullOrEmpty(s),
-                "Name is required.");
-
-            var minimumLengthErrorMessage = $"Minimum length is {minimumLength}";
-            var secondValidation = new BasePropertyValidation<TestViewModel, string>(
-                viewModel,
-                vm => vm.Name,
-                s => s.Length > minimumLength,
-                s => minimumLengthErrorMessage);
-
-            // Add validations
-            viewModel.ValidationContext.Add(firstValidation);
-            viewModel.ValidationContext.Add(secondValidation);
-
-            // View bindings
-            view.Bind(view.ViewModel, vm => vm.Name, v => v.NameLabel);
-
-            // View validations bindings
-            view.BindValidation(view.ViewModel, vm => vm.Name, v => v.NameErrorLabel);
-
-            viewModel.Name = "som";
-
-            Assert.False(viewModel.ValidationContext.IsValid);
-            Assert.Equal(2, viewModel.ValidationContext.Validations.Count);
-
-            // Checks if second validation error message is shown
-            Assert.Equal(minimumLengthErrorMessage, view.NameErrorLabel);
-        }
-
-        /// <summary>
-        /// Verifies that two validations properties are correctly applied
-        /// in a View property given by a complex expression.
-        /// </summary>
-        [Fact]
-        public void ShouldSupportBindingTwoValidationsForOnePropertyToChainedViewProperties()
-        {
-            const int minimumLength = 5;
-            var minimumLengthErrorMessage = $"Minimum length is {minimumLength}";
-            var viewModel = new TestViewModel { Name = "some" };
-            var view = new TestView(viewModel);
-
-            // Define the validations.
-            viewModel.ValidationRule(
-                vm => vm.Name,
-                s => !string.IsNullOrEmpty(s),
-                "Name is required.");
-
-            viewModel.ValidationRule(
-                vm => vm.Name,
-                s => s.Length > minimumLength,
-                minimumLengthErrorMessage);
-
-            // Define view bindings.
-            view.Bind(view.ViewModel, vm => vm.Name, v => v.NameLabel);
-            view.BindValidation(view.ViewModel, vm => vm.Name, v => v.NameErrorContainer.Text);
-
-            Assert.False(viewModel.ValidationContext.IsValid);
-            Assert.Equal(2, viewModel.ValidationContext.Validations.Count);
-            Assert.Equal(minimumLengthErrorMessage, view.NameErrorContainer.Text);
-        }
-
-        /// <summary>
-        /// Verifies that validations registered with different lambda names are retrieved successfully.
-        /// </summary>
-        [Fact]
-        public void RegisterValidationsWithDifferentLambdaNameWorksTest()
-        {
-            const string validName = "valid";
-            var viewModel = new TestViewModel { Name = validName };
-            var view = new TestView(viewModel);
-
-            var firstValidation = new BasePropertyValidation<TestViewModel, string>(
-                viewModel,
-                viewModelProperty => viewModelProperty.Name,
-                s => !string.IsNullOrEmpty(s),
-                s => $"Name {s} isn't valid");
-
-            // Add validations
-            viewModel.ValidationContext.Add(firstValidation);
-
-            // View bindings
-            view.Bind(view.ViewModel, vm => vm.Name, v => v.NameLabel);
-
-            // This was throwing exception due to naming problems with lambdas expressions
-            view.BindValidation(
-                view.ViewModel,
-                vm => vm.Name,
-                v => v.NameErrorLabel);
-
-            Assert.True(viewModel.ValidationContext.IsValid);
-            Assert.Single(viewModel.ValidationContext.Validations);
-        }
-
-        /// <summary>
-        /// Verifies that validation error messages get concatenated using white space.
-        /// </summary>
-        [Fact]
-        public void ValidationMessagesDefaultConcatenationTest()
-        {
-            var viewModel = new TestViewModel { Name = string.Empty };
-            var view = new TestView(viewModel);
-
-            var nameValidation = new BasePropertyValidation<TestViewModel, string>(
-                viewModel,
-                viewModelProperty => viewModelProperty.Name,
-                s => !string.IsNullOrEmpty(s),
-                "Name should not be empty.");
-
-            var name2Validation = new BasePropertyValidation<TestViewModel, string>(
-                viewModel,
-                viewModelProperty => viewModelProperty.Name2,
-                s => !string.IsNullOrEmpty(s),
-                "Name2 should not be empty.");
-
-            viewModel.ValidationContext.Add(nameValidation);
-            viewModel.ValidationContext.Add(name2Validation);
-
-            view.Bind(view.ViewModel, vm => vm.Name, v => v.NameLabel);
-            view.Bind(view.ViewModel, vm => vm.Name2, v => v.Name2Label);
-            view.BindValidation(view.ViewModel, v => v.NameErrorLabel);
-
-            Assert.False(viewModel.ValidationContext.IsValid);
-            Assert.Equal(2, viewModel.ValidationContext.Validations.Count);
-            Assert.NotEmpty(view.NameErrorLabel);
-            Assert.Equal("Name should not be empty. Name2 should not be empty.", view.NameErrorLabel);
-        }
-
-        /// <summary>
         /// Verifies that this.IsValid() extension method observes a
         /// <see cref="ValidationContext"/> and emits new values.
         /// </summary>
